A glimpse of a world without Xor and XorT
This is a half-baked implementation of typelevel/cats#1192.  Several
things are commented out because we don't yet use kind projector.
EitherOps should be a value class for performance.

This would fit much more nicely with `fs2.Task`, preventing any
conversions between `Xor` and `Either` for `Task.async` and `.attempt`
results.  Without this, we're probably looking at `Task.asyncXor` and
`task.attemptXor` syntax, which will come at the cost of extra
conversions.

The `widen` is irritating.  `Left` and `Right` have two type parameters
in the standard library, versus their one in cats.  This seems to come
up regularly in pattern matching.  Maybe most of it is hidden in
`EitherOps` in practice?

The greater irritation is the fight between type classes and `EitherOps`
for names like `.map` and `.flatMap`.  One has to be very careful with
imports.  This is discussed a bit on Gitter:
https://gitter.im/typelevel/cats?at=57a52c88483751d50f2d62fd